chore(cli): improvements and restructure
