+++ packages/cli/src/commands/assets/pull/index.test.ts
@@ -447,7 +447,8 @@ describe("assets pull command", () => {
 
     expect(assetFileExists(privateAsset)).toBeFalsy();
     const logFile = getLogFileContents(LOG_PREFIX);
-    expect(logFile).toContain("Error fetching data from the API");
+    expect(logFile).toContain("The user is not authorized to access the API");
+    expect(logFile).toContain('"httpCode":401');
     expect(logFile).toContain('"fetchAssets":{"total":1,"succeeded":0,"failed":1}');
     expect(logFile).toContain('"save":{"total":0,"succeeded":0,"failed":0}');
     expect(process.exitCode).toBe(1);
