+++ packages/cli/src/commands/stories/find/types.ts
@@ -1,4 +1,7 @@
-import type { Story } from "../constants";
+import type { StoriesQueryParams, Story } from "../constants";
+import type { UI } from "../../../lib/ui";
+import type { Logger } from "../../../lib/logger/logger";
+import type { Reporter } from "../../../lib/reporter/reporter";
 import type { PublishStatus } from "./filters";
 
 export type EntryType = "all" | "story" | "folder";
@@ -21,3 +24,13 @@ export interface FindOptions {
 }
 
 export type ClientFilter = (story: Story) => boolean;
+
+/** The modules and space scope every mode of the command runs against. */
+export type FindContext = {
+  spaceId: string;
+  params: StoriesQueryParams;
+  ui: UI;
+  logger: Logger;
+  reporter: Reporter;
+  verbose: boolean;
+};
